// dotnet/openapi/model/applicantModelListPagedResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApplicantModel } from './applicantModel';


export interface ApplicantModelListPagedResponse { 
    message?: string | null;
    statusCode?: number;
    data?: Array<ApplicantModel> | null;
    pageIndex?: number;
    pageSize?: number;
    totalRecords?: number;
}